+++ src/App.tsx
@@ -4,9 +4,9 @@ import Login from './screens/Login';
 import CustomerMain from './screens/Customers/CustomerMain';
 import CustomerProfile from './screens/Customers/CustomerProfile';
 import CustomerRecords from './screens/Customers/CustomerRecords';
-import AddCustomer from './screens/AddCustomer';
-import EditCustomer from './screens/EditCustomer';
-import AddMeterReading from './screens/AddMeterReading';
+import AddCustomer from './screens/Customers/AddCustomer';
+import EditCustomer from './screens/Customers/EditCustomer';
+import AddMeterReading from './screens/Customers/AddMeterReading';
 import Inventory from './screens/Inventory';
 import Incidents from './screens/Incidents';
 import Maintenance from './screens/Maintenance';
