+++ src/components/Checkbox.tsx
@@ -8,7 +8,6 @@ const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     root: {
       display: 'flex',
-      alignItems: 'center',
     },
     checkIcon: {
       width: '20px',
@@ -25,7 +24,7 @@ const useStyles = makeStyles((theme: Theme) =>
       borderRadius: '5px',
       width: '76px',
       height: '30px',
-      marginLeft: '-7px',
+      margin: '0.25rem 0 0 -7px',
       padding: '0px 10px',
     },
   }),
