+++ src/screens/Customers/CustomerRecords.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Typography } from '@material-ui/core';
 import BaseHeader from '../../components/BaseComponents/BaseHeader';
-import TabMenu from '../../components/TabMenu';
+import TabMenu from '../../components/CustomerRecords/TabMenu';
 import { withStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { RouteComponentProps } from 'react-router-dom';
 import { InvoiceRecord, MeterReadingRecord } from '../../utils/airtable/interface';
