+++ src/screens/Customers/EditCustomer.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { Typography } from '@material-ui/core';
-import BaseHeader from '../components/BaseComponents/BaseHeader';
-import TextField from '../components/TextField';
-import Checkbox from '../components/Checkbox';
-import Button from '../components/Button';
+import BaseHeader from '../../components/BaseComponents/BaseHeader';
+import TextField from '../../components/TextField';
+import Checkbox from '../../components/Checkbox';
+import Button from '../../components/Button';
 import { withStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Link, RouteComponentProps } from 'react-router-dom';
 
@@ -21,14 +20,10 @@ const styles = (theme: Theme) =>
       textAlign: 'left',
       color: theme.palette.text.primary,
     },
-    label: {
-      marginTop: '-55px',
-      marginBottom: '40px',
-    },
   });
 
 interface EditCustomerProps extends RouteComponentProps {
-  classes: { root: string; header: string; content: string; label: string; };
+  classes: { root: string; header: string; content: string };
   location: any;
 }
 
@@ -41,12 +36,12 @@ function EditCustomer(props: EditCustomerProps) {
         <BaseHeader title="Edit Customer" leftIcon="backNav" />
       </div>
       <form noValidate className={classes.content}>
-        <TextField label={"Name"} id={"name"} primary={true} />
-        <Checkbox label={"Select if customer is inactive"} />
-        <Checkbox label={"Select if customer has meter"} />
-        <TextField label={"Tariff Plan"} id={"tarrif-plan"} primary={true} />
-        <TextField label={"Reason"} id={"reason"} primary={true} />
-        <Button label={"SAVE"} />
+        <TextField label={'Name'} id={'name'} primary={true} />
+        <Checkbox label={'Select if customer is inactive'} />
+        <Checkbox label={'Select if customer has meter'} />
+        <TextField label={'Tariff Plan'} id={'tarrif-plan'} primary={true} />
+        <TextField label={'Reason'} id={'reason'} primary={true} />
+        <Button label={'SAVE'} />
       </form>
     </div>
   );
